update readme and some file organization
