+++ FlexNetworking/FlexNetworking+RxSwift.swift
@@ -24,11 +24,11 @@ extension FlexNetworking.Rx {
     /// Creates a Single observable corresponding to the result of a FlexNetworking request where pre-request and post-request hooks are omitted.
     /// When the returned Single is disposed, it cancels the task corresponding to the request initiated by the observable.
     ///
-    public func runRequestWithoutHooks(urlSession session: URLSession = .shared, path: String, method: String, body: RequestBody?, headers: [String: String] = [:]) -> Single<Response> {
+    public func runRequestWithoutHooks(session: URLSession = .shared, path: String, method: String, body: RequestBody?, headers: [String: String] = [:]) -> Single<Response> {
         return Single<Response>.create(subscribe: { observer in
             do {
                 let task = try self.flex.getTaskForRequest(
-                    urlSession: session,
+                    session: session,
                     path: path,
                     method: method,
                     body: body,
@@ -62,7 +62,7 @@ extension FlexNetworking.Rx {
     /// Creates a Single observable corresponding to the result of a FlexNetworking request.
     /// When the returned Single is disposed, it cancels the task corresponding to the request initiated by the observable.
     ///
-    public func runRequest(urlSession session: URLSession = .shared, path: String, method: String, body: RequestBody?, headers: [String: String] = [:]) -> Single<Response> {
+    public func runRequest(session: URLSession = .shared, path: String, method: String, body: RequestBody?, headers: [String: String] = [:]) -> Single<Response> {
         let getFinalRequestParameters = Single<RequestParameters>.create(subscribe: { observer in
             do {
                 let startingRequestParameters: RequestParameters = (session, path, method, body, headers)
@@ -79,7 +79,7 @@ extension FlexNetworking.Rx {
 
         let getInitialResponse = getFinalRequestParameters.flatMap { finalRequestParameters -> Single<(Response, RequestParameters)> in
             return self.runRequestWithoutHooks(
-                urlSession: finalRequestParameters.urlSession,
+                session: finalRequestParameters.session,
                 path: finalRequestParameters.path,
                 method: finalRequestParameters.method,
                 body: finalRequestParameters.body,
@@ -102,8 +102,8 @@ extension FlexNetworking.Rx {
                     return Single.just((response, originalRequestParameters, false))
                 case .continue:
                     return Single.just((response, originalRequestParameters, true))
-                case .makeNewRequest(let (urlSession, path, method, body, headers)):
-                    let nextResponse = self.runRequestWithoutHooks(urlSession: urlSession, path: path, method: method, body: body, headers: headers)
+                case .makeNewRequest(let (session, path, method, body, headers)):
+                    let nextResponse = self.runRequestWithoutHooks(session: session, path: path, method: method, body: body, headers: headers)
                     return nextResponse.map { response in (response, originalRequestParameters, true) }
                 }
             }
@@ -117,29 +117,6 @@ extension FlexNetworking.Rx {
             .subscribeOn(ConcurrentDispatchQueueScheduler(queue: DispatchQueue.global(qos: .default)))
     }
 
-    public struct DecodingError: Error, CustomNSError {
-        public let outputDTOTypeName: String
-        public let error: Error
-        public let response: Response
-
-        public static var errorDomain: String {
-            return "DecodingError"
-        }
-
-        public var errorCode: Int {
-            return 1
-        }
-
-        public var errorUserInfo: [String : Any] {
-            return [
-                "outputDTOTypeName": outputDTOTypeName,
-                "underlyingError": String(describing: error),
-                "underlyingErrorDescription": error.localizedDescription,
-                "response": String(describing: response)
-            ]
-        }
-    }
-
     ///
     /// Creates a Single observable corresponding to the result of a FlexNetworking request where the request body is specified as a JSON-encoded Codable.
     /// When the returned Single is disposed, it cancels the task corresponding to the request initiated by the observable.
@@ -150,7 +127,7 @@ extension FlexNetworking.Rx {
     /// If you specify a non-JSONEncoder instance for `encoder`, please ensure you specify a corresponding `contentType` as well, as the `Content-Type` header will otherwise be overwritten by `application/json`.
     ///
     public func requestCodable<InputDTO: Encodable, OutputDTO: Decodable>(
-        urlSession session: URLSession = .shared,
+        session: URLSession = .shared,
         encoder: JSONEncoder? = nil,
         contentType: String = "application/json",
         decoder: JSONDecoder? = nil,
@@ -173,7 +150,7 @@ extension FlexNetworking.Rx {
 
         let response = data.flatMap({ body -> Single<Response> in
             let body: RequestBody = RawBody(data: body, contentType: "application/json")
-            return self.runRequest(urlSession: session, path: path, method: method, body: body, headers: headers)
+            return self.runRequest(session: session, path: path, method: method, body: body, headers: headers)
         })
 
         let output = response.flatMap({ response -> Single<OutputDTO> in
@@ -196,7 +173,7 @@ extension FlexNetworking.Rx {
     /// If nil, or omitted (equivalent), the default decoder from the parent FlexNetworking instance will simply be used.
     ///
     public func requestCodable<OutputDTO: Decodable>(
-        urlSession session: URLSession = .shared,
+        session: URLSession = .shared,
         decoder: JSONDecoder? = nil,
         path: String,
         method: String,
@@ -206,7 +183,7 @@ extension FlexNetworking.Rx {
 
         let usableDecoder = decoder ?? self.flex.defaultDecoder
 
-        let response = self.runRequest(urlSession: session, path: path, method: method, body: body, headers: headers)
+        let response = self.runRequest(session: session, path: path, method: method, body: body, headers: headers)
 
         let output = response.flatMap({ response -> Single<OutputDTO> in
             do {
