+++ FlexNetworking/FlexNetworking.swift
@@ -6,236 +6,8 @@
 
 import Foundation
 
-public enum Result<T>: CustomStringConvertible {
-    case success(T)
-    case failure(Error)
-    
-    public var description: String {
-        switch self {
-        case .success(let value):
-            return "Success(\(value))"
-        case .failure(let error):
-            return "Failure(\(String(reflecting: error)))"
-        }
-    }
-}
-
-public typealias ResultBlock = (Result<Response>) -> ()
-
-public enum RequestError: Error, CustomNSError {
-    case noInternet(Error)
-    case miscURLSessionError(Error)
-    case invalidURL(message: String)
-    case emptyResponseError(Response)
-    case cancelledByCaller
-    case unknownError(message: String)
-
-    public var localizedDescription: String {
-        switch self {
-        case .noInternet(let error):
-            return "No internet (underlying error: \(error))"
-        case .miscURLSessionError(let error):
-            return "Miscellaneous URL session error. Underlying error code: \((error as NSError).code)\n\(error.localizedDescription)"
-        case .invalidURL(let message):
-            return "Invalid URL. Message: \(message)"
-        case .emptyResponseError(let response):
-            return "Response had no data; status was \(response.status)"
-        case .cancelledByCaller:
-            return "Request was cancelled by the caller"
-        case .unknownError(let message):
-            return "Unknown error. Message: \(message)"
-        }
-    }
-
-    public static var errorDomain: String {
-        return "RequestError"
-    }
-
-    public var errorCode: Int {
-        return 1
-    }
-
-    public var errorUserInfo: [String : Any] {
-        return ["description": self.localizedDescription]
-    }
-}
-
-public protocol RequestBody {
-    /// Called for GET requests.
-    func getQueryString() -> String?
-    
-    /// Called for POST requests.
-    func getHTTPBody() -> Data?
-    
-    /// Called for POST requests only.
-    func getContentType() -> String
-}
-
-let disallowedCharactersSet = CharacterSet(charactersIn: "!*'();:@&=+$,/?%#[] <>")
-extension String {
-    internal func urlEncoded() -> String? {
-        return self.addingPercentEncoding(withAllowedCharacters: disallowedCharactersSet.inverted)
-    }
-}
-
-extension Dictionary {
-    internal func getSerialization() -> String {
-        var serialization: [String] = []
-        for (key, value) in self {
-            let mirror = Mirror(reflecting: value)
-            var valueString = "\(value)"
-            if let displayStyle = mirror.displayStyle {
-                switch displayStyle {
-                case .optional:
-                    if let some = mirror.children.first {
-                        valueString = "\(some.value)"
-                    }
-                default: break
-                }
-            }
-            
-            guard let encodedKey = "\(key)".urlEncoded(), let encodedValue = valueString.urlEncoded() else {
-                assert(false, "Error URLEncoding key or value.")
-                continue
-            }
-            
-            serialization.append("\(encodedKey)=\(encodedValue)")
-        }
-
-        return serialization.joined(separator: "&")
-    }
-}
-
-extension Dictionary: RequestBody {
-    public func getQueryString() -> String? {
-        return self.getSerialization()
-    }
-    
-    public func getHTTPBody() -> Data? {
-        return self.getSerialization().data(using: .utf8)
-    }
-    
-    public func getContentType() -> String {
-        return "application/x-www-form-urlencoded"
-    }
-}
-
-@available (*, deprecated, message: "Pass Dictionary instance directly instead of wrapping it in a DictionaryBody() constructor.")
-public struct DictionaryBody: RequestBody {
-    private let queryDict: [String: Any]
-    
-    public init(_ queryDict: [String: Any]) {
-        self.queryDict = queryDict
-    }
-    
-    public func getQueryString() -> String? {
-        return self.queryDict.getSerialization()
-    }
-    
-    public func getHTTPBody() -> Data? {
-        return self.getQueryString()?.data(using: .utf8)
-    }
-    
-    public func getContentType() -> String {
-        return queryDict.getContentType()
-    }
-}
-
-/// RawBody allows you to specify data bytes with a given content-type, so you can pass any request body (incl. multipart form data).
-/// NB: On a GET request, a RawData body WILL be ignored.
-public struct RawBody: RequestBody {
-    public let data: Data
-    public let contentType: String
-    
-    public init(data: Data, contentType: String) {
-        self.data = data
-        self.contentType = contentType
-    }
-    
-    public func getQueryString() -> String? {
-        assert(false, "Do not use RawData with a GET request")
-        return nil
-    }
-    
-    public func getHTTPBody() -> Data? {
-        return self.data
-    }
-    
-    public func getContentType() -> String {
-        return self.contentType
-    }
-}
-
-public struct Response: CustomStringConvertible {
-    public let status: Int
-    
-    public let rawData: Data?
-    public let asString: String?
-
-    public let requestParameters: RequestParameters
-    
-    public var description: String {
-        let bodyDescription: String
-        if let string = self.asString {
-            bodyDescription = string
-        } else if let data = self.rawData {
-            bodyDescription = "\(data.count) bytes"
-        } else {
-            bodyDescription = "(null body)"
-        }
-        
-        return "Response(status=\(self.status)):\n\(bodyDescription)"
-    }
-}
-
-public typealias RequestParameters = (urlSession: URLSession, path: String, method: String, body: RequestBody?, headers: [String: String])
-
-public protocol PreRequestHook {
-    func execute(on requestParameters: RequestParameters) throws -> RequestParameters
-}
-
-public struct BlockPreRequestHook: PreRequestHook {
-    public typealias Block = (RequestParameters) throws -> RequestParameters
-
-    public let block: Block
-
-    public init(block: @escaping Block) {
-        self.block = block
-    }
-
-    public func execute(on requestParameters: RequestParameters) throws -> RequestParameters {
-        return try block(requestParameters)
-    }
-}
-
-public enum PostRequestHookResult {
-    /// Continues to the next hook in the chain, passing the unmodified last response as input to it.
-    case `continue`
-
-    /// Skips the rest of the chain, passing the current response to whatever completion handler was specified on the initial request.
-    case completed
-
-    /// Makes a new request, and passes the result of that request to the next post-request hook.
-    case makeNewRequest(RequestParameters)
-}
-
-public protocol PostRequestHook {
-    func execute(lastResponse: Response, originalRequestParameters: RequestParameters) throws -> PostRequestHookResult
-}
-
-public struct BlockPostRequestHook: PostRequestHook {
-    public typealias Block = (_ lastResponse: Response, _ originalRequestParameters: RequestParameters) throws -> PostRequestHookResult
-
-    public let block: Block
-
-    public init(block: @escaping Block) {
-        self.block = block
-    }
-
-    public func execute(lastResponse: Response, originalRequestParameters: RequestParameters) throws -> PostRequestHookResult {
-        return try block(lastResponse, originalRequestParameters)
-    }
-}
+/// All the context that goes into making a request. Useful for troubleshooting, as this is included with all responses.
+public typealias RequestParameters = (session: URLSession, path: String, method: String, body: RequestBody?, headers: [String: String])
 
 public class FlexNetworking {
     public let preRequestHooks: [PreRequestHook]
@@ -262,7 +34,7 @@ public class FlexNetworking {
     ///
     /// Creates a URLSessionTask for a single HTTP request with request parameters specified in FlexNetworking notation.
     ///
-    public func getTaskForRequest(urlSession session: URLSession, path: String, method: String, body: RequestBody?, headers: [String: String] = [:], completionHandler: @escaping (Data?, URLResponse?, Error?) -> Void) throws -> URLSessionTask {
+    public func getTaskForRequest(session: URLSession, path: String, method: String, body: RequestBody?, headers: [String: String] = [:], completionHandler: @escaping (Data?, URLResponse?, Error?) -> Void) throws -> URLSessionTask {
         guard let url = URL(string: path) else {
             throw RequestError.invalidURL(message: "Invalid URL \(path)")
         }
@@ -346,14 +118,14 @@ public class FlexNetworking {
     /// Token refresh can be implemented via `postRequestHooks`.
     /// Custom functionality that does not fit in `preRequestHooks` or `postRequestHooks` can be implemented by creating an extension of FlexNetworking with methods adopting the signatures you need, and delegating to internal FlexNetworking methods from those extension methods.
     ///
-    public func runRequest(urlSession session: URLSession, path: String, method: String, body: RequestBody?, headers: [String: String] = [:]) throws -> Response {
+    public func runRequest(session: URLSession = .shared, path: String, method: String, body: RequestBody?, headers: [String: String] = [:]) throws -> Response {
         let startingRequestParameters: RequestParameters = (session, path, method, body, headers)
         let finalRequestParameters = try preRequestHooks.reduce(startingRequestParameters) { (requestParameters, hook) -> RequestParameters in
             return try hook.execute(on: startingRequestParameters)
         }
 
         let initialResponse = try self.runRequestWithoutHooks(
-            urlSession: finalRequestParameters.urlSession,
+            session: finalRequestParameters.session,
             path: finalRequestParameters.path,
             method: finalRequestParameters.method,
             body: finalRequestParameters.body,
@@ -371,8 +143,8 @@ public class FlexNetworking {
                 breakLoop = true
             case .continue:
                 continue
-            case .makeNewRequest(let (urlSession, path, method, body, headers)):
-                lastResponse = try self.runRequestWithoutHooks(urlSession: urlSession, path: path, method: method, body: body, headers: headers)
+            case .makeNewRequest(let (session, path, method, body, headers)):
+                lastResponse = try self.runRequestWithoutHooks(session: session, path: path, method: method, body: body, headers: headers)
             }
 
             if breakLoop {
@@ -407,7 +179,7 @@ public class FlexNetworking {
     ///
     /// Returns a `Response` or throws a `RequestError`.
     ///
-    public func runRequestWithoutHooks(urlSession session: URLSession, path: String, method: String, body: RequestBody?, headers: [String: String] = [:]) throws -> Response {
+    public func runRequestWithoutHooks(session: URLSession, path: String, method: String, body: RequestBody?, headers: [String: String] = [:]) throws -> Response {
         let sema = DispatchSemaphore(value: 0)
 
         var httpURLResponse: HTTPURLResponse?
@@ -415,7 +187,7 @@ public class FlexNetworking {
         var responseData: Data?
 
         let task = try self.getTaskForRequest(
-            urlSession: session,
+            session: session,
             path: path,
             method: method,
             body: body,
@@ -467,10 +239,10 @@ public class FlexNetworking {
     /// Custom functionality that does not fit in `preRequestHooks` or `postRequestHooks` can be implemented by creating an extension of FlexNetworking with methods adopting the signatures you need, and delegating to internal FlexNetworking methods from those extension methods.
     /// If you need a cancel mechanism, look into the 'FlexNetworking/RxSwift' subpod.
     ///
-    public func runRequestAsync(_ urlSession: URLSession, path: String, method: String, body: RequestBody?, headers: [String: String] = [:], completion: ResultBlock?) {
+    public func runRequestAsync(session: URLSession, path: String, method: String, body: RequestBody?, headers: [String: String] = [:], completion: ResultBlock?) {
         DispatchQueue.global(qos: .userInitiated).async {
             do {
-                let response = try self.runRequest(urlSession: urlSession, path: path, method: method, body: body)
+                let response = try self.runRequest(session: session, path: path, method: method, body: body)
                 DispatchQueue.main.async {
                     completion?(.success(response))
                 }
